Added | CRUD Events

+++ 09-calendar-app/src/hooks/useCalendarStore.js
@@ -1,34 +1,69 @@
 import { useDispatch, useSelector } from 'react-redux';
-import { onAddNewEvent, onDeleteEvent, onSetActiveEvent, onUpdateEvent } from '../store';
+import Swal from 'sweetalert2';
+import calendarApi from '../api/calendarApi';
+import { convertEventsToDateEvents } from '../helpers';
+import { onAddNewEvent, onDeleteEvent, onLoadEvents, onSetActiveEvent, onUpdateEvent } from '../store';
 
 
 export const useCalendarStore = () => {
-  
     const dispatch = useDispatch();
     const { events, activeEvent } = useSelector( state => state.calendar );
+
+    const { user } = useSelector( state => state.auth );
 
     const setActiveEvent = ( calendarEvent ) => {
         dispatch( onSetActiveEvent( calendarEvent ) )
     }
 
     const startSavingEvent = async( calendarEvent ) => {
-        // TODO: llegar al backend
 
-        // Todo bien
-        if( calendarEvent._id ) {
+      try {
+        if( calendarEvent.id ) {
             // Actualizando
-            dispatch( onUpdateEvent({ ...calendarEvent }) );
-        } else {
-            // Creando
-            dispatch( onAddNewEvent({ ...calendarEvent, _id: new Date().getTime() }) );
+            await calendarApi.put(`/events/${ calendarEvent.id}`, calendarEvent )
+            dispatch( onUpdateEvent({ ...calendarEvent, user }) );
+            return;
         }
+
+        const { data } = await calendarApi.post('/events', calendarEvent );
+        dispatch( onAddNewEvent({ ...calendarEvent, id: data.event.id, user }) );
+
+      } catch (error) {
+        console.log(error);
+        Swal.fire('Error al guardar', error.response.data.msg, 'error');
+      }
+
     }
 
-    const startDeletingEvent = () => {
+    const startDeletingEvent = async() => {
         // Todo: Llegar al backend
+        try {
+          if( activeEvent.id ) {
+              // Borrando
+              await calendarApi.delete(`/events/${ activeEvent.id}`);
+              dispatch( onDeleteEvent() );
+              return;
+          }
+
+        } catch (error) {
+          console.log(error);
+          Swal.fire('Error al eliminar', error.response.data.msg, 'error');
+        }
 
 
-        dispatch( onDeleteEvent() );
+    }
+
+    const startLoadingEvent = async() => {
+      try {
+        const { data } = await calendarApi.get('/events');
+        const events = convertEventsToDateEvents(data.events);
+        dispatch( onLoadEvents( events ) );
+
+      } catch (error) {
+        console.log('Error al cargar eventos');
+        console.log(error);
+
+      }
     }
 
 
@@ -42,5 +77,6 @@ export const useCalendarStore = () => {
         startDeletingEvent,
         setActiveEvent,
         startSavingEvent,
+        startLoadingEvent
     }
 }
